[Tool] Rework simulation time period to be sync with actual stimuli

+++ openfpga/src/fpga_verilog/verilog_formal_random_top_testbench.cpp
@@ -250,10 +250,10 @@ void print_verilog_random_top_testbench(const std::string& circuit_name,
                                 clock_port_names,
                                 std::string(DEFAULT_CLOCK_NAME));
 
-  int simulation_time = find_operating_phase_simulation_time(MAGIC_NUMBER_FOR_SIMULATION_TIME,
-                                                             simulation_parameters.num_clock_cycles(),
-                                                             1./simulation_parameters.operating_clock_frequency(),
-                                                             VERILOG_SIM_TIMESCALE);
+  float simulation_time = find_operating_phase_simulation_time(MAGIC_NUMBER_FOR_SIMULATION_TIME,
+                                                               simulation_parameters.num_clock_cycles(),
+                                                               1./simulation_parameters.operating_clock_frequency(),
+                                                               VERILOG_SIM_TIMESCALE);
 
   /* Add Icarus requirement */
   print_verilog_timeout_and_vcd(fp,
